+++ renderer/shaders/skybox/skybox.frag
@@ -1,7 +1,7 @@
 layout(location = 0) out vec4 FragColor;
 layout(location = 1) out vec4 FragPosition;
 layout(location = 2) out vec4 FragNormal;
-layout(location = 3) out vec4 FragIDs;
+layout(location = 3) out uvec4 FragIDs;
 
 in VertexData {
     flat uint drawID;
@@ -16,5 +16,5 @@ void main() {
     FragColor = vec4(envColor, 1.0);
     FragPosition = vec4(fsIn.WorldPos, 1.0);
     FragNormal = vec4(0.0/0.0, 0.0/0.0, 0.0/0.0, 1.0); // make NaN
-    FragIDs = vec4(fsIn.drawID, gl_PrimitiveID, 0.0, 1.0);
+    FragIDs = uvec4(fsIn.drawID, gl_PrimitiveID, 0.0, 1.0);
 }
